+++ unity/Runtime/VoxtaChatSession.cs
@@ -1,5 +1,7 @@
 using System;
-using Voxta.Unity.Protocol.Generated;
+using Voxta.Model.Shared;
+using Voxta.Model.WebsocketMessages.ClientMessages;
+using Voxta.Model.WebsocketMessages.ServerMessages;
 
 namespace Voxta.Unity
 {