+++ unity/Tests/Runtime/VoxtaSpeechPlayerTests.cs
@@ -7,7 +7,8 @@ using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.Networking;
 using UnityEngine.TestTools;
-using Voxta.Unity.Protocol.Generated;
+using Voxta.Model.WebsocketMessages.ClientMessages;
+using Voxta.Model.WebsocketMessages.ServerMessages;
 using Voxta.Unity.Transport;
 
 namespace Voxta.Unity.Tests